+++ packages/hono-takibi/src/generator/zod/z/integer.ts
@@ -1,5 +1,5 @@
 import type { Schema } from '../../../openapi/types.js'
-import { _default, max, min, regex } from '../utils/index.js'
+import { _default, max, min, regex } from '../../../utils/index.js'
 
 /**
  * Generate a Zod integer schema string from an OpenAPI schema.
