+++ packages/hono-takibi/src/generator/zod/z/string.ts
@@ -1,5 +1,5 @@
 import type { Schema } from '../../../openapi/types.js'
-import { _default, max, min, regex } from '../utils/index.js'
+import { _default, max, min, regex } from '../../../utils/index.js'
 
 const FORMAT_STRING: Record<string, string> = {
   email: '.email()',
